+++ activerecord/lib/active_record/normalization.rb
@@ -32,7 +32,7 @@ module ActiveRecord # :nodoc:
       # Declares a normalization for one or more attributes. The normalization
       # is applied when the attribute is assigned or updated, and the normalized
       # value will be persisted to the database. The normalization is also
-      # applied to the corresponding keyword argument of finder methods. This
+      # applied to the corresponding keyword argument of query methods. This
       # allows a record to be created and later queried using unnormalized
       # values.
       #
@@ -69,7 +69,8 @@ module ActiveRecord # :nodoc:
       #   user.email                  # => "cruise-control@example.com"
       #   user.email_before_type_cast # => "cruise-control@example.com"
       #
-      #   User.where(email: "\tCRUISE-CONTROL@EXAMPLE.COM ").count # => 1
+      #   User.where(email: "\tCRUISE-CONTROL@EXAMPLE.COM ").count         # => 1
+      #   User.where(["email = ?", "\tCRUISE-CONTROL@EXAMPLE.COM "]).count # => 0
       #
       #   User.exists?(email: "\tCRUISE-CONTROL@EXAMPLE.COM ")         # => true
       #   User.exists?(["email = ?", "\tCRUISE-CONTROL@EXAMPLE.COM "]) # => false